fixed height bug and moved to using js modules
