+++ script.js
@@ -1,17 +1,11 @@
-//time display
+import { dateTime } from "./modules/date-time.js";
+
+//time display initialise
 const initDateTime = new Date().toLocaleTimeString(navigator.language, {
   hour: "2-digit",
   minute: "2-digit",
 });
 document.querySelector("#time").textContent = initDateTime;
 
-const dateTime = () => {
-  const now = new Date();
-  const currentDateTime = now.toLocaleTimeString(navigator.language, {
-    hour: "2-digit",
-    minute: "2-digit",
-  });
-  document.querySelector("#time").textContent = currentDateTime;
-};
 dateTime;
-setInterval(dateTime, 1000);
+setInterval(dateTime, 10000);
