+++ styles.scss
@@ -1,11 +1,13 @@
 @import "./partials/apps.scss";
 
+html,
 body {
   margin: 0;
 }
 .windows {
+  height: 100vh;
   &_desktop {
-    height: calc(95vh - 3px);
+    height: calc(95vh - 15px);
     background-image: url(./images/windows-xp-bliss-bg.jpg);
     display: flex;
     flex-direction: column;
